+++ WhatsAppAPISolutionDL/Dto/APIMessageDto.cs
@@ -9,11 +9,11 @@ namespace WhatsAppAPISolutionDL.Dto
     public partial class APIMessageDto
     {
         public int? ClientId { get; set; }
-        public long? SenderNameId { get; set; }
+        public int? SenderNameId { get; set; }
         public string TrxType { get; set; }
         public string Udf1 { get; set; }
         public string Udf2 { get; set; }
-        public long? TemplateId { get; set; }
+        public int? TemplateId { get; set; }
         public string PhoneNumber { get; set; }
         public string Url { get; set; }
         public int? Status { get; set; }
