+++ WhatsAppAPISolutionDL/Dto/RoleDto.cs
@@ -9,7 +9,7 @@ namespace WhatsAppAPISolutionDL.Dto
 {
     public partial class RoleDto
     {
-        public long? RoleId { get; set; }
+        public int? RoleId { get; set; }
         [Required]
         public int? ClientId { get; set; }
         [Required]
