+++ WhatsAppAPISolutionDL/Models/Page.cs
@@ -5,9 +5,9 @@ namespace WhatsAppAPISolutionDL.Models
 {
     public partial class Page
     {
-        public long PageId { get; set; }
+        public int PageId { get; set; }
         public int? ClientId { get; set; }
-        public long? PageName { get; set; }
+        public int? PageName { get; set; }
         public int? CreatedBy { get; set; }
         public DateTime? CreatedDate { get; set; }
         public int? UpdatedBy { get; set; }
