+++ WhatsAppAPISolutionDL/UserModels/UGroup.cs
@@ -8,8 +8,8 @@ namespace WhatsAppAPISolutionDL.UserModels
 {
     public partial class UGroup
     {
-        public long GroupId { get; set; }
-        public long ClientId { get; set; }
+        public int GroupId { get; set; }
+        public int ClientId { get; set; }
         public string GroupName { get; set; }
         public int? CreatedBy { get; set; }
         public string CreatedDate { get; set; }
