+++ WhatsAppAPISolutionDL/UserModels/USenderName.cs
@@ -8,7 +8,7 @@ namespace WhatsAppAPISolutionDL.UserModels
 {
     public partial class USenderName
     {
-        public long SenderId { get; set; }
+        public int SenderId { get; set; }
         public int? ClientId { get; set; }
         public string ClientName { get; set; }
         public string SenderName { get; set; }
