+++ WhatsAppAPISolutionDL/UserModels/UTemplateParameter.cs
@@ -8,7 +8,7 @@ namespace WhatsAppAPISolutionDL.UserModels
 {
     public partial class UTemplateParameter
     {
-        public long ParamId { get; set; }
+        public int ParamId { get; set; }
         public string TemplateId { get; set; }
         public string TemplateName { get; set; }
         public int? ClientId { get; set; }
